requirements.txt
albumentations==1.0.3
ipywidgets
opencv-python==4.5.5.64
Pillow==9.1.0
pandas==1.3.4
pycocotools==2.0.4
seaborn==0.11.2
segmentation-models-pytorch==0.2.0
scipy==1.7.3
webcolors==1.11.1

wandb
